+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/js/MessageBridgeRatDispatcherSpec.kt
@@ -47,6 +47,7 @@ class MessageBridgeRatDispatcherSpec {
         ActionType.CLOSE_MINIAPP.action to Actype.CLOSE_MINIAPP,
         ActionType.GET_HOST_APP_THEME_COLORS.action to Actype.GET_HOST_APP_THEME_COLORS,
         ActionType.GET_IS_DARK_MODE.action to Actype.GET_IS_DARK_MODE,
+        ActionType.UNIVERSAL_BRIDGE_INFO to Actype.RECEIVE_UNIVERSAL_BRIDGE_INFO,
         )
 
     @Test